+++ org.eclipse.gef/src/org/eclipse/gef/internal/ui/palette/editparts/DrawerEditPart.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2010 IBM Corporation and others.
+ * Copyright (c) 2000, 2024 IBM Corporation and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -65,7 +65,7 @@ public class DrawerEditPart extends PaletteEditPart implements IPinnableEditPart
 	 */
 	@Override
 	public IFigure createFigure() {
-		DrawerFigure fig = new DrawerFigure(getViewer().getControl()) {
+		DrawerFigure fig = new DrawerFigure(getViewer().getControl(), getColorPalette()) {
 			@Override
 			IFigure buildTooltip() {
 				return createToolTip();